+++ core/components/romanescobackyard/migxconfigs/content_related.config.js
@@ -8,7 +8,7 @@
                 {
                     "MIGX_id": 1,
                     "field": "source",
-                    "caption": "Page 1",
+                    "caption": "Source page",
                     "description": "",
                     "description_is_code": "0",
                     "inputTV": "",
@@ -35,7 +35,7 @@
                 {
                     "MIGX_id": 2,
                     "field": "destination",
-                    "caption": "Page 2",
+                    "caption": "Related page",
                     "description": "",
                     "description_is_code": "0",
                     "inputTV": "",
@@ -63,7 +63,7 @@
                     "MIGX_id": 3,
                     "field": "title",
                     "caption": "Title",
-                    "description": "",
+                    "description": "Briefly summarize the content of the related page, from the perspective of the source page. How does the story continue on the related page?",
                     "description_is_code": "0",
                     "inputTV": "",
                     "inputTVtype": "text",
@@ -82,7 +82,7 @@
                     "MIGX_id": 4,
                     "field": "description",
                     "caption": "Description",
-                    "description": "",
+                    "description": "Here you can expand on the title a bit more, if necessary.",
                     "description_is_code": "0",
                     "inputTV": "",
                     "inputTVtype": "textarea",
@@ -101,17 +101,19 @@
                     "MIGX_id": 5,
                     "field": "weight",
                     "caption": "Weight",
-                    "description": "",
+                    "description": "Indicate how closely the source and destination pages are related. More weight means a higher position in related content lists.",
                     "description_is_code": "0",
                     "inputTV": "",
-                    "inputTVtype": "number",
+                    "inputTVtype": "option",
                     "validation": "",
-                    "configs": "",
+                    "configs": {
+                        "columns": "20"
+                    },
                     "restrictive_condition": "",
                     "display": "",
                     "sourceFrom": "config",
                     "sources": "",
-                    "inputOptionValues": "",
+                    "inputOptionValues": "@CHUNK tvSelectFibonacci",
                     "default": "",
                     "useDefaultIfEmpty": "0",
                     "pos": 5
@@ -158,7 +160,7 @@
                     "MIGX_id": 98,
                     "field": "crosslink_id",
                     "caption": "Crosslink ID",
-                    "description": "",
+                    "description": "This is the ID of the MIGXdb object. For internal use only.",
                     "description_is_code": "0",
                     "inputTV": "",
                     "inputTVtype": "number",
@@ -233,8 +235,16 @@
         "check_resid_TV": "",
         "join_alias": "",
         "has_jointable": "no",
-        "getlistwhere": "{\"source\":[[+resource_id]]}",
-        "joins": "",
+        "getlistwhere": "",
+        "joins": [
+            {
+                "alias": "Source"
+            },{
+                "alias": "Destination"
+            },{
+                "alias": "Author"
+            }
+        ],
         "hooksnippets": {
             "aftersave": "migxSaveRelated"
         },
@@ -249,7 +259,7 @@
     "columns": [
         {
             "MIGX_id": 1,
-            "header": "Page 1",
+            "header": "Source",
             "dataIndex": "source",
             "width": 80,
             "sortable": true,
@@ -258,13 +268,13 @@
             "renderer": "this.renderChunk",
             "clickaction": "",
             "selectorconfig": "",
-            "renderchunktpl": "[[#[[+source]].pagetitle]]",
+            "renderchunktpl": "[[+Source_pagetitle]] ([[+source]])",
             "renderoptions": "",
             "editor": ""
         },
         {
             "MIGX_id": 2,
-            "header": "Page 2",
+            "header": "Destination",
             "dataIndex": "destination",
             "width": 80,
             "sortable": true,
@@ -273,7 +283,7 @@
             "renderer": "this.renderChunk",
             "clickaction": "",
             "selectorconfig": "",
-            "renderchunktpl": "[[#[[+destination]].pagetitle]]",
+            "renderchunktpl": "[[+Destination_pagetitle]] ([[+destination]])",
             "renderoptions": "",
             "editor": ""
         },
@@ -306,6 +316,36 @@
             "renderchunktpl": "",
             "renderoptions": "",
             "editor": "this.textEditor"
+        },
+        {
+            "MIGX_id": 12,
+            "header": "Weight",
+            "dataIndex": "weight",
+            "width": 20,
+            "sortable": true,
+            "show_in_grid": 1,
+            "customrenderer": "",
+            "renderer": "",
+            "clickaction": "",
+            "selectorconfig": "",
+            "renderchunktpl": "",
+            "renderoptions": "",
+            "editor": "this.textEditor"
+        },
+        {
+            "MIGX_id": 13,
+            "header": "Author",
+            "dataIndex": "Author_username",
+            "width": 30,
+            "sortable": true,
+            "show_in_grid": 1,
+            "customrenderer": "",
+            "renderer": "",
+            "clickaction": "",
+            "selectorconfig": "",
+            "renderchunktpl": "",
+            "renderoptions": "",
+            "editor": ""
         },
         {
             "MIGX_id": 98,